Another set of cleanup and improvements
Extracted all structs/enums in separate files
Separate filehelper in test app
