+++ Source/L2NextCrypto/Public/Crypto/Enums/DecryptErrorStatus.h
@@ -1,0 +1,12 @@
+#ifndef H_DECRYPT_ERROR_STATUS
+#define H_DECRYPT_ERROR_STATUS
+
+enum class EDecryptErrorStatus
+{
+	VERSION_NOT_SUPPORTED,
+	INVALID_HEADER,
+	INFLATE_FAILED,
+	NONE
+};
+
+#endif // H_DECRYPT_ERROR_STATUS
