+++ Source/L2NextCrypto/Public/Crypto/Enums/EncryptErrorStatus.h
@@ -1,0 +1,12 @@
+#ifndef H_ENCRYPT_ERROR_STATUS
+#define H_ENCRYPT_ERROR_STATUS
+
+enum class EEncryptErrorStatus
+{
+	VERSION_NOT_SUPPORTED,
+	INVALID_HEADER,
+	DEFLATE_FAILED,
+	NONE
+};
+
+#endif // H_ENCRYPT_ERROR_STATUS
