+++ Source/L2NextCrypto/Public/Crypto/Structs/DecryptResult.h
@@ -1,0 +1,14 @@
+
+#ifndef H_DECRYPT_RESULT
+#define H_DECRYPT_RESULT
+
+#include "Shared/Structs/FileData.h"
+
+struct SDecryptResult
+{
+	int version;
+	SFileData content;
+};
+
+
+#endif // H_DECRYPT_RESULT
