+++ Source/L2NextCrypto/Public/Crypto/Structs/EncryptResult.h
@@ -1,0 +1,13 @@
+#ifndef H_ENCRYPT_RESULT
+#define H_ENCRYPT_RESULT
+
+#include "Shared/Structs/FileData.h"
+
+struct SEncryptResult
+{
+	int version;
+	SFileData content;
+};
+
+
+#endif // H_ENCRYPT_RESULT
