+++ cmake/Library/Sources.cmake
@@ -37,11 +37,18 @@ add_source_group(Source.Crypto.Validators.HeaderValidator.Structs
 	"Public/Crypto/Validators/HeaderValidator/Structs/ValidationResult.h"
 )
 
-
 # Crypto enums.
 add_source_group(Source.Crypto.Enums
 	"Public/Crypto/Enums/CryptType.h"
 	"Public/Crypto/Enums/HeaderVersion.h"
+	"Public/Crypto/Enums/DecryptErrorStatus.h"
+	"Public/Crypto/Enums/EncryptErrorStatus.h"
+)
+
+# Crypto structs.
+add_source_group(Source.Crypto.Structs
+	"Public/Crypto/Structs/DecryptResult.h"
+	"Public/Crypto/Structs/EncryptResult.h"
 )
 
 # Crypto algorithms.
@@ -106,9 +113,13 @@ add_source_group(Source.Crypto.Algorithms.41x.Shared
 	"Public/Crypto/Algorithms/41x/Shared/Ver41xParams.h"
 )
 
-# Common.
+# Shared.
 add_source_group(Source.Shared.Templates
 	"Public/Shared/Templates/EventTemplate.h"
+)
+
+add_source_group(Source.Shared.Structs
+	"Public/Shared/Structs/FileData.h"
 )
 
 # Utils.
